remove document/spec concept
show dropdown

remove spec route and merged with default debug page

showing document as collection

remove document term in UI

fix lint issue

fix ut failures

replace api collection view name in ui
