+++ packages/insomnia/src/routes/organization.$organizationId.project.$projectId.workspace.$workspaceId.spec.tsx
@@ -1,1486 +1,17 @@
-import { type IRuleResult } from '@stoplight/spectral-core';
-import CodeMirror from 'codemirror';
-import { models, services } from 'insomnia-data';
-import type { OpenAPIV3 } from 'openapi-types';
-import { type ReactNode, useCallback, useEffect, useLayoutEffect, useMemo, useRef, useState } from 'react';
-import {
-  Button,
-  Dialog,
-  GridList,
-  GridListItem,
-  Heading,
-  ListBox,
-  ListBoxItem,
-  Menu,
-  MenuItem,
-  MenuTrigger,
-  Modal,
-  ModalOverlay,
-  Popover,
-  ToggleButton,
-  Tooltip,
-  TooltipTrigger,
-} from 'react-aria-components';
-import {
-  type ImperativePanelGroupHandle,
-  type ImperativePanelHandle,
-  Panel,
-  PanelGroup,
-  PanelResizeHandle,
-} from 'react-resizable-panels';
-import { href, redirect, useLoaderData } from 'react-router';
-import * as reactUse from 'react-use';
-import { SwaggerUIBundle } from 'swagger-ui-dist';
-import YAML from 'yaml';
-
-import { parseApiSpec } from '~/common/api-specs';
-import { DEFAULT_SIDEBAR_SIZE } from '~/common/constants';
-import { debounce } from '~/common/misc';
-import { utf8ByteLength } from '~/common/utils/utf8-bytes';
-import { useRootLoaderData } from '~/root';
-import { useDeleteProjectRulesetActionFetcher } from '~/routes/organization.$organizationId.project.$projectId.delete-ruleset';
-import { useRefreshProjectRulesetActionFetcher } from '~/routes/organization.$organizationId.project.$projectId.refresh-ruleset';
-import { useUpdateProjectRulesetActionFetcher } from '~/routes/organization.$organizationId.project.$projectId.update-ruleset';
-import {
-  useWorkspaceLoaderData,
-  WORKSPACE_CONTENT_WRAPPER,
-} from '~/routes/organization.$organizationId.project.$projectId.workspace.$workspaceId';
-import { useSpecGenerateRequestCollectionActionFetcher } from '~/routes/organization.$organizationId.project.$projectId.workspace.$workspaceId.spec.generate-request-collection';
-import { useSpecUpdateActionFetcher } from '~/routes/organization.$organizationId.project.$projectId.workspace.$workspaceId.spec.update';
-import { AnalyticsEvent } from '~/ui/analytics';
-import { CodeEditor, type CodeEditorHandle } from '~/ui/components/.client/codemirror/code-editor';
-import { Badge } from '~/ui/components/base/badge';
-import { DesignEmptyState } from '~/ui/components/design-empty-state';
-import { DocumentTab } from '~/ui/components/document-tab';
-import { Icon } from '~/ui/components/icon';
-import { useDocBodyKeyboardShortcuts } from '~/ui/components/keydown-binder';
-import { showError, showModal } from '~/ui/components/modals';
-import { AskModal } from '~/ui/components/modals/ask-modal';
-import { CookiesModal } from '~/ui/components/modals/cookies-modal';
-import { NewWorkspaceModal } from '~/ui/components/modals/new-workspace-modal';
-import { CertificatesModal } from '~/ui/components/modals/workspace-certificates-modal';
-import { WorkspaceEnvironmentsEditModal } from '~/ui/components/modals/workspace-environments-edit-modal';
-import { OrganizationTabList } from '~/ui/components/tabs/tab-list';
-import { formatMethodName } from '~/ui/components/tags/method-tag';
-import { showResourceNotFoundToast, showToast } from '~/ui/components/toast-notification';
-import WorkspacePaneHeader from '~/ui/components/workspace/workspace-pane-header';
-import { useAIFeatureStatus } from '~/ui/hooks/use-organization-features';
-import { useOrganizationStorageRule } from '~/ui/hooks/use-organization-storage-rule';
-import { useGitVCSVersion } from '~/ui/hooks/use-vcs-version';
-import { resolveGitRepoBaseDir } from '~/ui/utils/git-repo-path';
-import { selectFileOrFolder } from '~/ui/utils/select-file-or-folder';
+import { href, redirect } from 'react-router';
 
 import type { Route } from './+types/organization.$organizationId.project.$projectId.workspace.$workspaceId.spec';
 
+// The spec route has been merged into the debug route.
+// This route only exists to redirect old bookmarks/tabs/synced links to the merged page.
 export async function clientLoader({ params }: Route.ClientLoaderArgs) {
   const { organizationId, projectId, workspaceId } = params;
 
-  const project = await services.project.getById(projectId);
-  if (!project) {
-    showResourceNotFoundToast(`Project not found: ${projectId}`);
-    throw redirect(href('/organization/:organizationId/project', { organizationId }));
-  }
-
-  const workspace = await services.workspace.getById(workspaceId);
-  if (!workspace) {
-    showResourceNotFoundToast(`Workspace not found: ${workspaceId}`);
-    throw redirect(href('/organization/:organizationId/project/:projectId', { organizationId, projectId }));
-  }
-
-  const apiSpec = await services.apiSpec.getByParentId(workspaceId);
-  if (!apiSpec) {
-    showResourceNotFoundToast(`API Specification not found for workspace: ${workspaceId}`);
-    throw redirect(href('/organization/:organizationId/project/:projectId', { organizationId, projectId }));
-  }
-
-  const workspaceMeta = await services.workspaceMeta.getByParentId(workspaceId);
-  const isConnectedGitProject = models.project.isConnectedGitProject(project);
-
-  const gitRepositoryId = isConnectedGitProject
-    ? models.project.getEffectiveRepoId(project)
-    : workspaceMeta?.gitRepositoryId;
-  // we don't run the lint here because it is expensive and slows first render too much
-  // TODO: add this in once we run this loader outside the renderer
-  // Honour user-chosen repo locations: the ruleset lives at the repo root, which
-  // is the GitRepository's `directory` when set, else the managed location.
-  const gitRepository = gitRepositoryId ? await services.gitRepository.getById(gitRepositoryId) : null;
-  const gitSyncRulesetPath = gitRepository
-    ? window.path.join(resolveGitRepoBaseDir(gitRepository), '.spectral.yaml')
-    : '';
-
-  // The ProjectLintRuleset record is the source of truth for both git and cloud projects.
-  // For git, the RepoFileWatcher keeps .spectral.yaml in sync with this record.
-  const projectLintRuleset = await services.projectLintRuleset.getByParentId(projectId);
-  const rulesetContent = projectLintRuleset?.rulesetContent || '';
-  const rulesetLastCompiledAt = projectLintRuleset?.modified ?? null;
-
-  // For git projects, surface a committed .spectral.yaml that was rejected as
-  // invalid on import (so the user learns it wasn't applied and why, instead of
-  // silently falling back to the default ruleset). Refreshed automatically via
-  // the git.db-synced → revalidate flow in root.tsx.
-  const rulesetImportIssue =
-    isConnectedGitProject && gitRepositoryId
-      ? await window.main.git.getProjectRulesetImportIssue({ projectId, gitRepositoryId })
-      : null;
-
-  let parsedSpec: OpenAPIV3.Document | undefined;
-
-  try {
-    parsedSpec = YAML.parse(apiSpec.contents) as OpenAPIV3.Document;
-  } catch {}
-
-  return {
-    apiSpec,
-    gitSyncRulesetPath,
-    isConnectedGitProject,
-    parsedSpec,
-    rulesetContent,
-    rulesetLastCompiledAt,
-    rulesetImportIssue,
-  };
+  throw redirect(
+    `${href('/organization/:organizationId/project/:projectId/workspace/:workspaceId/debug', {
+      organizationId,
+      projectId,
+      workspaceId,
+    })}`,
+  );
 }
-
-const SwaggerUIDiv = ({ text }: { text: string }) => {
-  useEffect(() => {
-    let spec = {};
-    try {
-      spec = parseApiSpec(text).contents || {};
-    } catch {}
-    SwaggerUIBundle({ spec, dom_id: '#swagger-ui' });
-  }, [text]);
-  return (
-    <div
-      id="swagger-ui"
-      style={{
-        overflowY: 'auto',
-        height: '100%',
-        background: '#FFF',
-      }}
-    />
-  );
-};
-
-interface LintMessage {
-  type: 'error' | 'warning' | 'info';
-  message: string;
-  line: number;
-  range: IRuleResult['range'];
-  code: string;
-  path: string;
-}
-interface GroupedLintMessage {
-  code: string;
-  type: 'error' | 'warning' | 'info';
-  message: string;
-  occurrences: { line: number; range: IRuleResult['range']; path: string }[];
-}
-
-interface SpecActionItem {
-  id: string;
-  name: string;
-  icon?: ReactNode;
-  badge?: ReactNode;
-  isDisabled?: boolean;
-  tooltip?: string;
-  action: () => void;
-}
-
-const getMethodsFromOpenApiPathItem = (pathItem: OpenAPIV3.PathItemObject): string[] => {
-  const methods = ['get', 'put', 'post', 'delete', 'options', 'head', 'patch', 'trace'].filter(
-    method =>
-      // @ts-expect-error -- shrug I don't care what pathItem has in it
-      pathItem[method],
-  );
-
-  return methods;
-};
-
-const lintOptions = {
-  delay: 1000,
-};
-
-const Component = ({ params }: Route.ComponentProps) => {
-  const { organizationId, projectId, workspaceId } = params;
-  const { activeProject, vcsVersion } = useWorkspaceLoaderData()!;
-  const { settings } = useRootLoaderData()!;
-
-  const [isCookieModalOpen, setIsCookieModalOpen] = useState(false);
-  const [isEnvironmentModalOpen, setEnvironmentModalOpen] = useState(false);
-  const [_isEnvironmentPickerOpen, setIsEnvironmentPickerOpen] = useState(false);
-  const [isCertificatesModalOpen, setCertificatesModalOpen] = useState(false);
-  const [isNewMockServerModalOpen, setNewMockServerModalOpen] = useState(false);
-  const [isViewRulesetModalOpen, setIsViewRulesetModalOpen] = useState(false);
-
-  const storageRules = useOrganizationStorageRule(organizationId);
-
-  const { isGenerateMockServersWithAIEnabled } = useAIFeatureStatus();
-
-  const {
-    apiSpec,
-    gitSyncRulesetPath,
-    isConnectedGitProject,
-    parsedSpec,
-    rulesetContent,
-    rulesetLastCompiledAt,
-    rulesetImportIssue,
-  } = useLoaderData<typeof clientLoader>();
-
-  const [lintMessages, setLintMessages] = useState<LintMessage[]>([]);
-  const [expandedLintMessageCodes, setExpandedLintMessageCodes] = useState<string[]>([]);
-
-  const editor = useRef<CodeEditorHandle>(null);
-  const { submit: updateApiSpec } = useSpecUpdateActionFetcher();
-  const { submit: updateProjectRuleset } = useUpdateProjectRulesetActionFetcher();
-  const { submit: deleteProjectRuleset } = useDeleteProjectRulesetActionFetcher();
-  const { submit: refreshProjectRuleset } = useRefreshProjectRulesetActionFetcher();
-  const [isRefreshing, setIsRefreshing] = useState(false);
-  const generateRequestCollectionFetcher = useSpecGenerateRequestCollectionActionFetcher();
-  const gitVersion = useGitVCSVersion();
-  const [isLintPaneOpen, setIsLintPaneOpen] = useState(false);
-  const [isSpecPaneOpen, setIsSpecPaneOpen] = useState(Boolean(parsedSpec));
-  const [selectedRulesetPath, setSelectedRulesetPath] = useState<string>('');
-
-  // Spectral requires a file path on disk to lint with a ruleset. Ref: lint-process.mjs.
-  // Cloud/local projects have no RepoFileWatcher, so rulesetContent from NeDB is mirrored
-  // to this per-project scratch path. Git projects lint against gitSyncRulesetPath, which
-  // the RepoFileWatcher keeps in sync with the record.
-  const rulesetWritePath = useMemo(
-    () => window.path.join(window.app.getPath('userData'), `projects/${projectId}/.spectral.yaml`),
-    [projectId],
-  );
-
-  const rulesetHasRemoteExtendsEntries = useMemo(() => {
-    if (!rulesetContent) return false;
-
-    try {
-      const parsedRuleset = YAML.parse(rulesetContent);
-      if (!parsedRuleset.extends) return false;
-
-      const formattedRulesetExtendsEntries = Array.isArray(parsedRuleset.extends)
-        ? parsedRuleset.extends
-        : [parsedRuleset.extends];
-      return formattedRulesetExtendsEntries.some((entry: string) => entry.startsWith('https://'));
-    } catch {
-      return false;
-    }
-  }, [rulesetContent]);
-
-  const groupedLintMessages = useMemo<GroupedLintMessage[]>(() => {
-    const map = new Map<string, GroupedLintMessage>();
-    for (const msg of lintMessages) {
-      const key = msg.code;
-      const existing = map.get(key);
-      if (existing) {
-        existing.occurrences.push({ line: msg.line, range: msg.range, path: msg.path });
-      } else {
-        map.set(key, {
-          code: msg.code ?? '',
-          type: msg.type,
-          message: msg.message,
-          occurrences: [{ line: msg.line, range: msg.range, path: msg.path }],
-        });
-      }
-    }
-
-    return Array.from(map.values());
-  }, [lintMessages]);
-
-  const { components, info, servers, paths } = parsedSpec || {};
-  const { requestBodies, responses, parameters, headers, schemas, securitySchemes } = components || {};
-
-  const lintErrors = lintMessages.filter(message => message.type === 'error');
-  const lintWarnings = lintMessages.filter(message => message.type === 'warning');
-
-  const registerCodeMirrorLint = useCallback(
-    (rulesetContent: string) => {
-      CodeMirror.registerHelper('lint', 'openapi', async (contents: string) => {
-        try {
-          const { diagnostics, error, cancelled } = await window.main.lintSpec({
-            documentContent: contents,
-            projectId,
-            rulesetContent,
-          });
-          if (cancelled) {
-            return [];
-          }
-          if (error) {
-            console.log('Handled error detected while linting:', error);
-            setLintMessages([]);
-            showError({
-              title: 'Linting Error',
-              message: `An error occurred while linting the OpenAPI specification: ${error}`,
-            });
-            return [];
-          }
-          const lintResult = diagnostics?.map(({ severity, code, message, range, path }) => {
-            return {
-              from: CodeMirror.Pos(range.start.line, range.start.character),
-              to: CodeMirror.Pos(range.end.line, range.end.character),
-              code: String(code),
-              message,
-              severity: ['error', 'warning'][severity] ?? 'info',
-              type: (['error', 'warning'][severity] ?? 'info') as LintMessage['type'],
-              range,
-              line: range.start.line,
-              path: path.join('.'),
-            };
-          });
-          setLintMessages?.(lintResult || []);
-          return lintResult;
-        } catch (error) {
-          // return a rejected promise so that codemirror do nothing
-          console.log('Unhandled error while linting:', error);
-          setLintMessages([]);
-          showError({
-            title: 'Linting Error',
-            message: `An error occurred while linting the OpenAPI specification: ${error}`,
-          });
-          return [];
-        }
-      });
-    },
-    [projectId],
-  );
-
-  useEffect(() => {
-    registerCodeMirrorLint(rulesetContent);
-    // when first time into document editor, the lint helper register later than codemirror init, we need to trigger lint through execute setOption
-    editor.current?.tryToSetOption('lint', { ...lintOptions });
-  }, [rulesetContent, projectId, registerCodeMirrorLint]);
-
-  useEffect(() => {
-    setSelectedRulesetPath(
-      isConnectedGitProject && gitSyncRulesetPath && rulesetContent
-        ? gitSyncRulesetPath
-        : rulesetContent
-          ? rulesetWritePath
-          : '',
-    );
-  }, [gitSyncRulesetPath, isConnectedGitProject, rulesetWritePath, rulesetContent]);
-
-  reactUse.useUnmount(() => {
-    // delete the helper to avoid it run multiple times when user enter the page next time
-    CodeMirror.registerHelper('lint', 'openapi', () => {});
-  });
-
-  const onCodeEditorChange = useMemo(() => {
-    const handler = async (contents: string) => {
-      return updateApiSpec({
-        organizationId,
-        projectId,
-        workspaceId,
-        contents: contents,
-      });
-    };
-
-    return debounce(handler, 500);
-  }, [organizationId, projectId, updateApiSpec, workspaceId]);
-
-  const handleScrollToSelection = useCallback(
-    (chStart: number, chEnd: number, lineStart: number, lineEnd: number) => {
-      if (!editor.current) {
-        return;
-      }
-      editor.current.scrollToSelection(chStart, chEnd, lineStart, lineEnd);
-    },
-    [editor],
-  );
-
-  const handleScrollToLintMessage = useCallback(
-    (occurrence: GroupedLintMessage['occurrences'][number]) => {
-      if (!editor.current) {
-        return;
-      }
-      const { start, end } = occurrence.range;
-      editor.current.scrollToSelection(start.character, end.character, start.line, end.line);
-    },
-    [editor],
-  );
-
-  const [expandedKeys, setExpandedKeys] = useState<string[]>([]);
-
-  const navigateToPath = (path: string): void => {
-    const pathSegments = path.split('.');
-    const scrollPosition = {
-      start: { line: 0, col: 0 },
-      end: { line: 0, col: 200 },
-    };
-
-    try {
-      JSON.parse(apiSpec.contents);
-      // Account for JSON (as string) line number shift
-      scrollPosition.start.line = 1;
-    } catch {}
-
-    const lineCounter = new YAML.LineCounter();
-    const doc = YAML.parseDocument(apiSpec.contents, { lineCounter });
-    const astNode = doc.getIn(pathSegments, true) as YAML.Node;
-    const nodePosition = astNode.range && lineCounter.linePos(astNode.range[0]);
-    if (nodePosition) {
-      scrollPosition.start.line += nodePosition.line;
-    }
-    const isServersSection = pathSegments[0] === 'servers';
-    if (!isServersSection) {
-      scrollPosition.start.line -= 1;
-    }
-
-    scrollPosition.end.line = scrollPosition.start.line;
-    // NOTE: We're subtracting 1 from everything because YAML CST uses
-    //   1-based indexing and we use 0-based.
-    handleScrollToSelection(
-      scrollPosition.start.col - 1,
-      scrollPosition.end.col - 1,
-      scrollPosition.start.line - 1,
-      scrollPosition.end.line - 1,
-    );
-  };
-
-  const sidebarPanelRef = useRef<ImperativePanelGroupHandle>(null);
-  const lintPanelRef = useRef<ImperativePanelHandle>(null);
-
-  useDocBodyKeyboardShortcuts({
-    environment_showEditor: () => setEnvironmentModalOpen(true),
-    environment_showSwitchMenu: () => setIsEnvironmentPickerOpen(true),
-    showCookiesEditor: () => setIsCookieModalOpen(true),
-  });
-
-  const specFormat = useMemo((): 'json' | 'yaml' | null => {
-    const contents = apiSpec.contents?.trim();
-    if (!contents) {
-      return null;
-    }
-    try {
-      JSON.parse(contents);
-      return 'json';
-    } catch {
-      return 'yaml';
-    }
-  }, [apiSpec.contents]);
-
-  const switchFormat = (to: 'json' | 'yaml') => {
-    const editorValue = editor.current?.getValue();
-    if (!editorValue) {
-      return;
-    }
-    let parsedSpec: string | undefined;
-    try {
-      // yaml parses json correctly
-      parsedSpec = YAML.parse(editorValue);
-    } catch {
-      showToast({
-        title: 'Failed to convert spec format',
-        icon: 'circle-exclamation',
-        status: 'error',
-        description: `Spec is not valid, cannot convert to ${to.toUpperCase()}`,
-      });
-      return;
-    }
-    const contents = to === 'json' ? JSON.stringify(parsedSpec, null, 2) : YAML.stringify(parsedSpec);
-    editor.current?.setValue(contents);
-    updateApiSpec({ organizationId, projectId, workspaceId, contents });
-  };
-
-  const handleSelectSpectralFile = async () => {
-    const { filePath, canceled } = await selectFileOrFolder({
-      itemTypes: ['file'],
-      extensions: ['yaml', 'yml'],
-      showHiddenFiles: true,
-    });
-
-    if (canceled || !filePath) {
-      return;
-    }
-
-    // We bundle the ruleset to resolve any extended rulesets and to validate the content
-    const { content, error } = await window.main.bundleSpectralRuleset({ sourcePath: filePath });
-    if (error || !content) {
-      showError({
-        title: 'Invalid Spectral Ruleset',
-        message: error ?? 'Failed to bundle ruleset.',
-      });
-      return;
-    }
-
-    const RULESET_MAX_BYTES = 1 * 1024 * 1024; // 1 MB
-    if (utf8ByteLength(content) > RULESET_MAX_BYTES) {
-      showError({
-        title: 'Ruleset Too Large',
-        message: 'The selected ruleset exceeds the maximum allowed size of 1 MB.',
-      });
-      return;
-    }
-
-    await updateProjectRuleset({ organizationId, projectId, rulesetContent: content });
-
-    window.main.trackAnalyticsEvent({
-      event: AnalyticsEvent.uploadLintRulesetClicked,
-      properties: {
-        project_type: models.project.isGitProject(activeProject)
-          ? 'git'
-          : models.project.isRemoteProject(activeProject)
-            ? 'remote'
-            : 'local',
-      },
-    });
-
-    setSelectedRulesetPath(gitSyncRulesetPath || rulesetWritePath);
-  };
-
-  const handleRefreshRuleset = async () => {
-    if (!rulesetContent) {
-      return;
-    }
-    setIsRefreshing(true);
-    try {
-      await window.main.refreshCompiledRuleset({ projectId, rulesetContent });
-      refreshProjectRuleset({ organizationId, projectId });
-      editor.current?.tryToSetOption('lint', { ...lintOptions });
-    } catch (err) {
-      showError({
-        title: 'Refresh Failed',
-        message: `Failed to refresh ruleset: ${err instanceof Error ? err.message : String(err)}`,
-      });
-    } finally {
-      setIsRefreshing(false);
-    }
-  };
-
-  const handleUnselectSpectralFile = async () => {
-    showModal(AskModal, {
-      title: 'Remove Ruleset File',
-      message:
-        'Are you sure you want to remove this custom ruleset? This will disable all custom linting rules and use the default Spectral ruleset.',
-      yesText: 'Remove',
-      color: 'danger',
-      noText: 'Cancel',
-      onDone: async (confirmed: boolean) => {
-        if (confirmed) {
-          await deleteProjectRuleset({
-            organizationId,
-            projectId,
-          });
-          await window.main.deleteCompiledRuleset({ projectId });
-          setSelectedRulesetPath('');
-        }
-      },
-    });
-  };
-
-  const generateActionList: SpecActionItem[] = [
-    {
-      id: 'generate-request-collection',
-      name: 'Collection',
-      icon: (
-        <span className="flex h-5 w-5 items-center justify-center rounded-sm bg-(--color-surprise) text-(--color-font-surprise)">
-          <Icon className="w-3" icon="bars" />
-        </span>
-      ),
-      isDisabled: !apiSpec.contents || lintErrors.length > 0 || generateRequestCollectionFetcher.state !== 'idle',
-      tooltip:
-        lintErrors.length > 0
-          ? 'You cannot generate a collection when spec errors exist. Fix the errors or change the ruleset first.'
-          : undefined,
-      action: () =>
-        generateRequestCollectionFetcher.submit({
-          organizationId,
-          projectId,
-          workspaceId,
-        }),
-    },
-    {
-      id: 'generate-mock-server',
-      name: 'Mock Server',
-      icon: (
-        <span className="flex h-5 w-5 items-center justify-center rounded-sm bg-(--color-warning) text-(--color-font-warning)">
-          <Icon className="w-3" icon="server" />
-        </span>
-      ),
-      badge: <Badge icon="sparkles" color="surprise" label="AI" style={{ marginLeft: 'auto', marginRight: '0' }} />,
-      isDisabled: !apiSpec.contents || !isGenerateMockServersWithAIEnabled,
-      action: () => {
-        window.main.trackAnalyticsEvent({
-          event: AnalyticsEvent.designerGenerateMockClicked,
-        });
-        setNewMockServerModalOpen(true);
-      },
-    },
-  ];
-
-  const specFormatActionList: SpecActionItem[] = [
-    {
-      id: 'json',
-      name: 'JSON',
-      action: () => switchFormat('json'),
-    },
-    {
-      id: 'yaml',
-      name: 'YAML',
-      action: () => switchFormat('yaml'),
-    },
-  ];
-
-  const generateDisabledKeys = generateActionList.filter(item => item.isDisabled).map(item => item.id);
-
-  const historyKey = `${apiSpec?._id}::${apiSpec?.created}::${gitVersion}::${vcsVersion}`;
-
-  const [direction, setDirection] = useState<'horizontal' | 'vertical'>(
-    settings.forceVerticalLayout ? 'vertical' : 'horizontal',
-  );
-  useLayoutEffect(() => {
-    if (settings.forceVerticalLayout) {
-      setDirection('vertical');
-      return () => {};
-    }
-    // Listen on media query changes
-    const mediaQuery = window.matchMedia('(max-width: 880px)');
-    setDirection(mediaQuery.matches ? 'vertical' : 'horizontal');
-
-    const handleChange = (e: MediaQueryListEvent) => {
-      setDirection(e.matches ? 'vertical' : 'horizontal');
-    };
-
-    mediaQuery.addEventListener('change', handleChange);
-
-    return () => {
-      mediaQuery.removeEventListener('change', handleChange);
-    };
-  }, [settings.forceVerticalLayout, direction]);
-
-  const specEditor = (
-    <div className="relative flex h-full w-full overflow-hidden">
-      <CodeEditor
-        id="spec-editor"
-        key={historyKey}
-        showPrettifyButton
-        ref={editor}
-        lintOptions={lintOptions}
-        mode={apiSpec.contents ? 'openapi' : undefined}
-        defaultValue={apiSpec.contents || ''}
-        onChange={onCodeEditorChange}
-        historyKey={historyKey}
-      />
-      {apiSpec.contents ? null : (
-        <DesignEmptyState
-          onImport={value => {
-            updateApiSpec({
-              organizationId,
-              projectId,
-              workspaceId,
-              contents: value,
-              fromTemplate: true,
-            });
-          }}
-        />
-      )}
-    </div>
-  );
-
-  const specPaneToolbar = apiSpec.contents ? (
-    <div className="flex h-[40px] shrink-0 items-center gap-2 overflow-hidden border-b border-solid border-(--hl-md) px-(--padding-sm)">
-      <span className="truncate text-sm text-(--hl) italic">
-        {parsedSpec?.openapi ? `OpenAPI ${parsedSpec.openapi}` : ''}
-      </span>
-      <span className="flex-1" />
-      <MenuTrigger>
-        <Button
-          aria-label="Generate"
-          className="flex shrink-0 items-center justify-center gap-2 rounded-md border border-solid border-(--hl-md) px-2.5 py-1 text-sm text-(--color-font) ring-1 ring-transparent transition-all hover:bg-(--hl-xs) focus:ring-(--hl-md) focus:ring-inset aria-pressed:bg-(--hl-sm)"
-        >
-          <span>Generate</span>
-          <Icon icon="chevron-down" className="w-2.5 text-(--hl)" />
-        </Button>
-        <Popover className="flex min-w-max flex-col overflow-y-hidden">
-          <Menu
-            aria-label="Generate menu"
-            selectionMode="single"
-            disabledKeys={generateDisabledKeys}
-            onAction={key => {
-              const item = generateActionList.find(item => item.id === key);
-              if (item) {
-                item.action();
-              }
-            }}
-            items={generateActionList}
-            className="min-w-max overflow-y-auto rounded-md border border-solid border-(--hl-sm) bg-(--color-bg) py-2 text-sm shadow-lg select-none focus:outline-hidden"
-          >
-            {item => (
-              <TooltipTrigger delay={0} isDisabled={!item.tooltip}>
-                <MenuItem
-                  id={item.id}
-                  textValue={item.name}
-                  className="flex h-(--line-height-xs) w-full items-center gap-2 bg-transparent px-(--padding-sm) whitespace-nowrap text-(--color-font) transition-colors hover:bg-(--hl-sm) focus:bg-(--hl-xs) focus:outline-hidden disabled:cursor-not-allowed aria-disabled:cursor-not-allowed aria-disabled:text-(--hl-md) aria-selected:font-bold"
-                  aria-label={item.name}
-                >
-                  {item.icon}
-                  <span>{item.name}</span>
-                  {item.badge && (
-                    <span className="flex origin-left scale-90 items-center pl-2 text-xs">{item.badge}</span>
-                  )}
-                </MenuItem>
-                <Tooltip
-                  offset={8}
-                  className="max-w-xs rounded-md border border-solid border-(--hl-sm) bg-(--color-bg) px-2 py-1 text-sm text-(--color-font) shadow-lg"
-                >
-                  {item.tooltip}
-                </Tooltip>
-              </TooltipTrigger>
-            )}
-          </Menu>
-        </Popover>
-      </MenuTrigger>
-      <MenuTrigger>
-        <Button
-          aria-label="Spec format"
-          className="flex shrink-0 items-center justify-center gap-2 rounded-md border border-solid border-(--hl-md) px-2.5 py-1 text-sm text-(--color-font) ring-1 ring-transparent transition-all hover:bg-(--hl-xs) focus:ring-(--hl-md) focus:ring-inset aria-pressed:bg-(--hl-sm)"
-        >
-          <span>{specFormat ? specFormat.toUpperCase() : 'JSON'}</span>
-          <Icon icon="chevron-down" className="w-2.5 text-(--hl)" />
-        </Button>
-        <Popover className="flex min-w-max flex-col overflow-y-hidden">
-          <div className="min-w-max overflow-y-auto rounded-md border border-solid border-(--hl-sm) bg-(--color-bg) py-2 text-sm shadow-lg select-none">
-            <span className="block px-(--padding-md) py-1 text-xs text-(--hl)">Spec format</span>
-            <Menu
-              aria-label="Spec format menu"
-              selectionMode="single"
-              selectedKeys={specFormat ? [specFormat] : []}
-              onAction={key => {
-                const item = specFormatActionList.find(item => item.id === key);
-                if (item) {
-                  item.action();
-                }
-              }}
-              items={specFormatActionList}
-              className="min-w-max focus:outline-hidden"
-            >
-              {item => (
-                <MenuItem
-                  className="flex h-(--line-height-xs) w-full items-center gap-2 bg-transparent px-(--padding-sm) whitespace-nowrap text-(--color-font) transition-colors hover:bg-(--hl-sm) focus:bg-(--hl-xs) focus:outline-hidden aria-selected:font-bold"
-                  aria-label={item.name}
-                >
-                  <span>{item.name}</span>
-                </MenuItem>
-              )}
-            </Menu>
-          </div>
-        </Popover>
-      </MenuTrigger>
-      <TooltipTrigger delay={0}>
-        <ToggleButton
-          aria-label="Toggle preview"
-          data-testid="preview-toggle"
-          isSelected={isSpecPaneOpen}
-          className="flex shrink-0 items-center justify-center rounded-md border border-solid border-(--hl-md) px-2 py-1 text-sm text-(--color-font) ring-1 ring-transparent transition-all hover:bg-(--hl-xs) focus:ring-(--hl-md) focus:ring-inset"
-          onChange={value => {
-            setIsSpecPaneOpen(value);
-            window.main.trackAnalyticsEvent({
-              event: AnalyticsEvent.designerPreviewToggled,
-              properties: {
-                status: !value ? 'open' : 'collapsed',
-              },
-            });
-          }}
-        >
-          {({ isSelected }) => (
-            <span className="flex h-5 w-5 items-center justify-center">
-              <Icon icon={isSelected ? 'eye-slash' : 'eye'} fixedWidth />
-            </span>
-          )}
-        </ToggleButton>
-        <Tooltip
-          offset={8}
-          className="rounded-md border border-solid border-(--hl-sm) bg-(--color-bg) px-2 py-1 text-sm text-(--color-font) shadow-lg"
-        >
-          {isSpecPaneOpen ? 'Hide docs preview' : 'Show docs preview'}
-        </Tooltip>
-      </TooltipTrigger>
-    </div>
-  ) : null;
-
-  const isRulesetInvalid = !!rulesetImportIssue && !rulesetContent;
-
-  const lintToolbar = (
-    <div
-      className={`flex h-(--line-height-sm) items-center gap-2 overflow-hidden border-solid border-(--hl-md) px-(--padding-sm) ${isLintPaneOpen ? 'border-b' : ''}`}
-    >
-      <div className="inline-flex items-center gap-2">
-        <Icon icon={selectedRulesetPath && !isRulesetInvalid ? 'file-circle-check' : 'file-circle-xmark'} />
-        {!isRulesetInvalid && selectedRulesetPath && (
-          <>
-            <TooltipTrigger delay={0}>
-              <Button
-                aria-label="View selected ruleset content"
-                className="underline"
-                onPress={() => setIsViewRulesetModalOpen(true)}
-              >
-                Custom Ruleset
-              </Button>
-              <Tooltip
-                placement="top end"
-                offset={8}
-                className="max-h-[85vh] max-w-xs overflow-y-auto rounded-md border border-solid border-(--hl-sm) bg-(--color-bg) px-4 py-2 text-sm text-(--color-font) shadow-lg select-none focus:outline-hidden"
-              >
-                <p className="mb-2">Using ruleset from</p>
-                <code className="block p-0 break-all whitespace-pre-wrap">{selectedRulesetPath}</code>
-              </Tooltip>
-            </TooltipTrigger>
-            {rulesetHasRemoteExtendsEntries && (
-              <TooltipTrigger delay={0}>
-                <Button
-                  aria-label="Refresh ruleset from remote sources"
-                  isDisabled={isRefreshing}
-                  onPress={handleRefreshRuleset}
-                  className="flex aspect-square h-6 shrink-0 items-center justify-center rounded-xs text-sm text-(--color-font) ring-1 ring-transparent transition-all hover:bg-(--hl-xs) focus:ring-(--hl-md) focus:ring-inset disabled:opacity-50 aria-pressed:bg-(--hl-sm)"
-                >
-                  <Icon icon={isRefreshing ? 'spinner' : 'rotate'} className={isRefreshing ? 'animate-spin' : ''} />
-                </Button>
-                <Tooltip
-                  placement="top end"
-                  offset={8}
-                  className="max-h-[85vh] max-w-xs overflow-y-auto rounded-md border border-solid border-(--hl-sm) bg-(--color-bg) px-4 py-2 text-sm text-(--color-font) shadow-lg select-none focus:outline-hidden"
-                >
-                  <p>Recompile ruleset, including re-fetching any referenced remote entries.</p>
-                  {rulesetLastCompiledAt && (
-                    <p className="mt-1">{`Last updated ${new Date(rulesetLastCompiledAt).toLocaleString()}`}.</p>
-                  )}
-                </Tooltip>
-              </TooltipTrigger>
-            )}
-            <TooltipTrigger delay={0}>
-              <Button
-                aria-label="Remove custom ruleset"
-                onPress={handleUnselectSpectralFile}
-                className="flex aspect-square h-6 shrink-0 items-center justify-center rounded-xs text-sm text-(--color-font) ring-1 ring-transparent transition-all hover:bg-(--hl-xs) focus:ring-(--hl-md) focus:ring-inset aria-pressed:bg-(--hl-sm)"
-              >
-                <Icon icon="xmark" />
-              </Button>
-              <Tooltip
-                placement="top end"
-                offset={8}
-                className="max-h-[85vh] max-w-xs overflow-y-auto rounded-md border border-solid border-(--hl-sm) bg-(--color-bg) px-4 py-2 text-sm text-(--color-font) shadow-lg select-none focus:outline-hidden"
-              >
-                <p>Clear custom ruleset and use default OAS ruleset</p>
-              </Tooltip>
-            </TooltipTrigger>
-          </>
-        )}
-        {!selectedRulesetPath && (
-          <>
-            <span>Default OAS Ruleset</span>
-            {isRulesetInvalid && (
-              <TooltipTrigger delay={0}>
-                <Button
-                  aria-label="Invalid ruleset details"
-                  className="flex aspect-square h-6 shrink-0 items-center justify-center rounded-xs text-sm text-(--color-font) ring-1 ring-transparent transition-all hover:bg-(--hl-xs) focus:ring-(--hl-md) focus:ring-inset aria-pressed:bg-(--hl-sm)"
-                >
-                  <Icon icon="triangle-exclamation" className="text-(--color-warning)" />
-                </Button>
-                <Tooltip
-                  placement="top end"
-                  offset={8}
-                  className="max-h-[85vh] max-w-xs overflow-y-auto rounded-md border border-solid border-(--hl-sm) bg-(--color-bg) px-4 py-2 text-sm text-(--color-font) shadow-lg select-none focus:outline-hidden"
-                >
-                  <p className="mb-2">
-                    The custom ruleset in this project contains invalid content: {rulesetImportIssue}
-                  </p>
-                  <p>Fix the syntax in your connected git repository. The default OAS ruleset is used until then.</p>
-                </Tooltip>
-              </TooltipTrigger>
-            )}
-            <TooltipTrigger delay={0}>
-              <Button
-                aria-label="Upload custom ruleset"
-                onPress={handleSelectSpectralFile}
-                className="flex aspect-square h-6 shrink-0 items-center justify-center rounded-xs text-sm text-(--color-font) ring-1 ring-transparent transition-all hover:bg-(--hl-xs) focus:ring-(--hl-md) focus:ring-inset aria-pressed:bg-(--hl-sm)"
-              >
-                <Icon icon="upload" />
-              </Button>
-              <Tooltip
-                placement="top end"
-                offset={8}
-                className="max-h-[85vh] max-w-xs overflow-y-auto rounded-md border border-solid border-(--hl-sm) bg-(--color-bg) px-4 py-2 text-sm text-(--color-font) shadow-lg select-none focus:outline-hidden"
-              >
-                <p>
-                  Upload a custom Spectral ruleset
-                  {isConnectedGitProject && (
-                    <span>
-                      {' '}
-                      or add a <code className="p-0">.spectral.yaml</code> file to the root of your connected git
-                      repository
-                    </span>
-                  )}
-                  . Any local files or remote URLs referenced via <code className="p-0">extends</code> will be bundled
-                  into a single ruleset on upload.
-                </p>
-              </Tooltip>
-            </TooltipTrigger>
-          </>
-        )}
-      </div>
-      <span className="flex-1" />
-      <div className="flex items-center gap-2">
-        {lintErrors.length > 0 && (
-          <div className="flex items-center gap-2 select-none">
-            <Icon icon="circle-xmark" className="text-(--color-danger)" />
-          </div>
-        )}
-        {lintWarnings.length > 0 && (
-          <div className="flex items-center gap-2 select-none">
-            <Icon icon="triangle-exclamation" className="text-(--color-warning)" />
-          </div>
-        )}
-        {apiSpec.contents && (
-          <div className="flex items-center gap-2 select-none">
-            {lintMessages.length === 0 && <Icon icon="check-square" className="text-(--color-success)" />}
-            {lintMessages.length === 0 ? (
-              'No lint problems'
-            ) : (
-              <Button
-                data-testid="lint-panel-toggle"
-                onPress={() => (isLintPaneOpen ? lintPanelRef.current?.collapse() : lintPanelRef.current?.expand())}
-              >
-                <span className="underline">
-                  {lintErrors.length} {lintErrors.length === 1 ? 'error' : 'errors'}, {lintWarnings.length}{' '}
-                  {lintWarnings.length === 1 ? 'warning' : 'warnings'}
-                </span>
-              </Button>
-            )}
-          </div>
-        )}
-      </div>
-    </div>
-  );
-
-  const lintMessageList = (
-    <div className="flex-1 overflow-y-auto select-none">
-      {groupedLintMessages.map(item => (
-        <div
-          key={item.code}
-          className="flex flex-col text-xs"
-          aria-expanded={expandedLintMessageCodes.includes(item.code)}
-        >
-          <Button
-            className="flex cursor-pointer items-center gap-2 p-(--padding-sm) outline-hidden transition-colors hover:bg-(--hl-sm)"
-            onPress={() =>
-              setExpandedLintMessageCodes(prev =>
-                prev.includes(item.code) ? prev.filter(c => c !== item.code) : [...prev, item.code],
-              )
-            }
-          >
-            <Icon
-              icon={expandedLintMessageCodes.includes(item.code) ? 'chevron-down' : 'chevron-right'}
-              className="h-2.5 w-2.5"
-            />
-            ({item.occurrences.length})
-            <Icon
-              className={item.type === 'error' ? 'text-(--color-danger)' : 'text-(--color-warning)'}
-              icon={item.type === 'error' ? 'circle-xmark' : 'triangle-exclamation'}
-            />
-            <span className="truncate">
-              {item.code}: {item.message}
-            </span>
-          </Button>
-          {expandedLintMessageCodes.includes(item.code) && (
-            <div className="flex flex-col gap-0.5">
-              {item.occurrences.map(occurrence => (
-                <Button
-                  key={`${occurrence.line}-${occurrence.path}-${occurrence.range.start.line}:${occurrence.range.start.character}-${occurrence.range.end.line}:${occurrence.range.end.character}`}
-                  className="flex gap-2 p-(--padding-sm) pl-[22px] text-left hover:bg-(--hl-sm)"
-                  onPress={() => handleScrollToLintMessage(occurrence)}
-                >
-                  <span className="shrink-0 underline">Ln {occurrence.line + 1}</span>
-                  {occurrence.path && <span className="truncate opacity-60">{occurrence.path}</span>}
-                </Button>
-              ))}
-            </div>
-          )}
-        </div>
-      ))}
-    </div>
-  );
-
-  return (
-    <div className="flex h-full flex-col">
-      <OrganizationTabList />
-      <WorkspacePaneHeader hasSettings={false} />
-      <PanelGroup
-        ref={sidebarPanelRef}
-        autoSaveId="insomnia-sidebar"
-        id={WORKSPACE_CONTENT_WRAPPER}
-        className="new-sidebar w-full flex-1 text-(--color-font)"
-        direction="horizontal"
-      >
-        <Panel
-          id="sidebar"
-          className="sidebar theme--sidebar"
-          defaultSize={DEFAULT_SIDEBAR_SIZE}
-          maxSize={40}
-          minSize={10}
-          collapsible
-        >
-          <div className="flex flex-1 flex-col divide-y divide-solid divide-(--hl-md) overflow-hidden">
-            <DocumentTab
-              organizationId={organizationId}
-              projectId={projectId}
-              workspaceId={workspaceId}
-              className="border-b border-solid border-(--hl-sm)"
-            />
-            <div className="flex flex-1 flex-col divide-y divide-solid divide-(--hl-md) overflow-y-auto">
-              {/* Info */}
-              {info && (
-                <div className="divide-y divide-solid divide-(--hl-md)">
-                  <Button
-                    className="flex w-full items-center justify-between gap-2 p-(--padding-sm) text-sm text-(--hl) uppercase select-none hover:bg-(--hl-sm) focus:bg-(--hl-sm)"
-                    onPress={() => {
-                      expandedKeys.includes('info')
-                        ? setExpandedKeys(expandedKeys.filter(key => key !== 'info'))
-                        : setExpandedKeys([...expandedKeys, 'info']);
-                    }}
-                  >
-                    <span className="truncate">Info</span>
-                    <Icon icon={expandedKeys.includes('info') ? 'minus' : 'plus'} className="text-xs" />
-                  </Button>
-                  {/* Info */}
-                  {expandedKeys.includes('info') && (
-                    <ListBox onAction={key => navigateToPath(key.toString())}>
-                      <ListBoxItem
-                        className="relative flex h-(--line-height-xs) w-full items-center gap-2 overflow-hidden px-4 text-(--hl) outline-hidden transition-colors select-none hover:bg-(--hl-xs) focus:bg-(--hl-sm)"
-                        id="info.title"
-                      >
-                        <span className="truncate">Title: {info.title}</span>
-                      </ListBoxItem>
-                      <ListBoxItem
-                        className="relative flex h-(--line-height-xs) w-full items-center gap-2 overflow-hidden px-4 text-(--hl) outline-hidden transition-colors select-none hover:bg-(--hl-xs) focus:bg-(--hl-sm)"
-                        id="info.description"
-                      >
-                        <span className="truncate">Description: {info.description}</span>
-                      </ListBoxItem>
-                      <ListBoxItem
-                        className="relative flex h-(--line-height-xs) w-full items-center gap-2 overflow-hidden px-4 text-(--hl) outline-hidden transition-colors select-none hover:bg-(--hl-xs) focus:bg-(--hl-sm)"
-                        id="info.version"
-                      >
-                        <span className="truncate">Version: {info.version}</span>
-                      </ListBoxItem>
-                      <ListBoxItem
-                        className="relative flex h-(--line-height-xs) w-full items-center gap-2 overflow-hidden px-4 text-(--hl) outline-hidden transition-colors select-none hover:bg-(--hl-xs) focus:bg-(--hl-sm)"
-                        id="info.license"
-                      >
-                        <span className="truncate">License: {info.license?.name}</span>
-                      </ListBoxItem>
-                    </ListBox>
-                  )}
-                </div>
-              )}
-              {/* Servers */}
-              {servers && (
-                <div className="divide-y divide-solid divide-(--hl-md)">
-                  <div>
-                    <Button
-                      className="flex w-full items-center justify-between gap-2 p-(--padding-sm) text-sm text-(--hl) uppercase select-none hover:bg-(--hl-sm) focus:bg-(--hl-sm)"
-                      onPress={() => {
-                        expandedKeys.includes('servers')
-                          ? setExpandedKeys(expandedKeys.filter(key => key !== 'servers'))
-                          : setExpandedKeys([...expandedKeys, 'servers']);
-                      }}
-                    >
-                      <span className="truncate">Servers</span>
-                      <Icon icon={expandedKeys.includes('servers') ? 'minus' : 'plus'} className="text-xs" />
-                    </Button>
-                  </div>
-                  {expandedKeys.includes('servers') && (
-                    <ListBox
-                      items={servers.map((server, index) => ({
-                        path: index,
-                        ...server,
-                      }))}
-                      onAction={key => navigateToPath(key.toString())}
-                    >
-                      {item => (
-                        <ListBoxItem
-                          className="relative flex h-(--line-height-xs) w-full items-center gap-2 overflow-hidden px-4 text-(--hl) outline-hidden transition-colors select-none hover:bg-(--hl-xs) focus:bg-(--hl-sm)"
-                          id={`servers.${item.path}`}
-                        >
-                          {item.url}
-                        </ListBoxItem>
-                      )}
-                    </ListBox>
-                  )}
-                </div>
-              )}
-              {/* Paths */}
-              {paths && (
-                <div className="divide-y divide-solid divide-(--hl-md)">
-                  <div>
-                    <Button
-                      className="flex w-full items-center justify-between gap-2 p-(--padding-sm) text-sm text-(--hl) uppercase select-none hover:bg-(--hl-sm) focus:bg-(--hl-sm)"
-                      onPress={() => {
-                        expandedKeys.includes('paths')
-                          ? setExpandedKeys(expandedKeys.filter(key => key !== 'paths'))
-                          : setExpandedKeys([...expandedKeys, 'paths']);
-                      }}
-                    >
-                      <span className="truncate">Paths</span>
-                      <Icon icon={expandedKeys.includes('paths') ? 'minus' : 'plus'} className="text-xs" />
-                    </Button>
-                  </div>
-                  {expandedKeys.includes('paths') && (
-                    <GridList
-                      items={Object.entries(paths).map(([path, item]) => ({
-                        ...item,
-                        id: path,
-                        path,
-                      }))}
-                      onAction={key => navigateToPath(key.toString())}
-                    >
-                      {item => (
-                        <GridListItem className="group outline-hidden select-none" id={`paths.${item.path}`}>
-                          <div className="relative flex h-(--line-height-xs) w-full items-center gap-2 overflow-hidden px-4 text-(--hl) outline-hidden transition-colors select-none group-hover:bg-(--hl-xs) group-focus:bg-(--hl-sm) group-aria-selected:text-(--color-font)">
-                            <span className="truncate">{item.path}</span>
-                            <span className="flex-1" />
-                            {getMethodsFromOpenApiPathItem(item).map(method => (
-                              <Button
-                                key={method}
-                                onPress={() => navigateToPath(`paths.${item.path}.${method}`)}
-                                className={`flex w-10 shrink-0 items-center justify-center rounded-xs border border-solid border-(--hl-sm) text-[0.65rem] http-method-${method.toUpperCase()}`}
-                              >
-                                {formatMethodName(method.toUpperCase())}
-                              </Button>
-                            ))}
-                          </div>
-                        </GridListItem>
-                      )}
-                    </GridList>
-                  )}
-                </div>
-              )}
-              {/* RequestBodies */}
-              {requestBodies && (
-                <div className="divide-y divide-solid divide-(--hl-md)">
-                  <div>
-                    <Button
-                      className="flex w-full items-center justify-between gap-2 p-(--padding-sm) text-sm text-(--hl) uppercase select-none hover:bg-(--hl-sm) focus:bg-(--hl-sm)"
-                      onPress={() => {
-                        expandedKeys.includes('requestBodies')
-                          ? setExpandedKeys(expandedKeys.filter(key => key !== 'requestBodies'))
-                          : setExpandedKeys([...expandedKeys, 'requestBodies']);
-                      }}
-                    >
-                      <span className="truncate">Request bodies</span>
-                      <Icon icon={expandedKeys.includes('requestBodies') ? 'minus' : 'plus'} className="text-xs" />
-                    </Button>
-                  </div>
-                  {expandedKeys.includes('requestBodies') && (
-                    <ListBox
-                      items={Object.entries(requestBodies).map(([path, item]) => ({
-                        ...item,
-                        id: path,
-                        path,
-                      }))}
-                      onAction={key => navigateToPath(key.toString())}
-                    >
-                      {item => (
-                        <ListBoxItem
-                          className="relative flex h-(--line-height-xs) w-full items-center gap-2 overflow-hidden px-4 text-(--hl) outline-hidden transition-colors select-none hover:bg-(--hl-xs) focus:bg-(--hl-sm)"
-                          id={`components.requestBodies.${item.path}`}
-                        >
-                          <span className="truncate">{item.path}</span>
-                        </ListBoxItem>
-                      )}
-                    </ListBox>
-                  )}
-                </div>
-              )}
-              {/* Responses */}
-              {responses && (
-                <div className="divide-y divide-solid divide-(--hl-md)">
-                  <div>
-                    <Button
-                      className="flex w-full items-center justify-between gap-2 p-(--padding-sm) text-sm text-(--hl) uppercase select-none hover:bg-(--hl-sm) focus:bg-(--hl-sm)"
-                      onPress={() => {
-                        expandedKeys.includes('responses')
-                          ? setExpandedKeys(expandedKeys.filter(key => key !== 'responses'))
-                          : setExpandedKeys([...expandedKeys, 'responses']);
-                      }}
-                    >
-                      <span className="truncate">Responses</span>
-                      <Icon icon={expandedKeys.includes('responses') ? 'minus' : 'plus'} className="text-xs" />
-                    </Button>
-                  </div>
-                  {expandedKeys.includes('responses') && (
-                    <ListBox
-                      items={Object.entries(responses).map(([path, item]) => ({
-                        ...item,
-                        id: path,
-                        path,
-                      }))}
-                      onAction={key => navigateToPath(key.toString())}
-                    >
-                      {item => (
-                        <ListBoxItem
-                          className="relative flex h-(--line-height-xs) w-full items-center gap-2 overflow-hidden px-4 text-(--hl) outline-hidden transition-colors select-none hover:bg-(--hl-xs) focus:bg-(--hl-sm)"
-                          id={`components.responses.${item.path}`}
-                        >
-                          <span className="truncate">{item.path}</span>
-                        </ListBoxItem>
-                      )}
-                    </ListBox>
-                  )}
-                </div>
-              )}
-              {/* Parameters */}
-              {parameters && (
-                <div className="divide-y divide-solid divide-(--hl-md)">
-                  <div>
-                    <Button
-                      className="flex w-full items-center justify-between gap-2 p-(--padding-sm) text-sm text-(--hl) uppercase select-none hover:bg-(--hl-sm) focus:bg-(--hl-sm)"
-                      onPress={() => {
-                        expandedKeys.includes('parameters')
-                          ? setExpandedKeys(expandedKeys.filter(key => key !== 'parameters'))
-                          : setExpandedKeys([...expandedKeys, 'parameters']);
-                      }}
-                    >
-                      <span className="truncate">Parameters</span>
-                      <Icon icon={expandedKeys.includes('parameters') ? 'minus' : 'plus'} className="text-xs" />
-                    </Button>
-                  </div>
-                  {expandedKeys.includes('parameters') && (
-                    <ListBox
-                      items={Object.entries(parameters).map(([path, item]) => ({
-                        ...item,
-                        id: path,
-                        path,
-                      }))}
-                      onAction={key => navigateToPath(key.toString())}
-                    >
-                      {item => (
-                        <ListBoxItem
-                          className="relative flex h-(--line-height-xs) w-full items-center gap-2 overflow-hidden px-4 text-(--hl) outline-hidden transition-colors select-none hover:bg-(--hl-xs) focus:bg-(--hl-sm)"
-                          id={`components.parameters.${item.path}`}
-                        >
-                          <span className="truncate">{item.path}</span>
-                        </ListBoxItem>
-                      )}
-                    </ListBox>
-                  )}
-                </div>
-              )}
-              {/* Headers */}
-              {headers && (
-                <div className="divide-y divide-solid divide-(--hl-md)">
-                  <div>
-                    <Button
-                      className="flex w-full items-center justify-between gap-2 p-(--padding-sm) text-sm text-(--hl) uppercase select-none hover:bg-(--hl-sm) focus:bg-(--hl-sm)"
-                      onPress={() => {
-                        expandedKeys.includes('headers')
-                          ? setExpandedKeys(expandedKeys.filter(key => key !== 'headers'))
-                          : setExpandedKeys([...expandedKeys, 'headers']);
-                      }}
-                    >
-                      <span className="truncate">Headers</span>
-                      <Icon icon={expandedKeys.includes('headers') ? 'minus' : 'plus'} className="text-xs" />
-                    </Button>
-                  </div>
-                  {expandedKeys.includes('headers') && (
-                    <ListBox
-                      items={Object.entries(headers).map(([path, item]) => ({
-                        ...item,
-                        id: path,
-                        path,
-                      }))}
-                      onAction={key => navigateToPath(key.toString())}
-                    >
-                      {item => (
-                        <ListBoxItem
-                          className="relative flex h-(--line-height-xs) w-full items-center gap-2 overflow-hidden px-4 text-(--hl) outline-hidden transition-colors select-none hover:bg-(--hl-xs) focus:bg-(--hl-sm)"
-                          id={`components.headers.${item.path}`}
-                        >
-                          <span className="truncate">{item.path}</span>
-                        </ListBoxItem>
-                      )}
-                    </ListBox>
-                  )}
-                </div>
-              )}
-              {/* Schemas */}
-              {schemas && (
-                <div className="divide-y divide-solid divide-(--hl-md)">
-                  <div>
-                    <Button
-                      className="flex w-full items-center justify-between gap-2 p-(--padding-sm) text-sm text-(--hl) uppercase select-none hover:bg-(--hl-sm) focus:bg-(--hl-sm)"
-                      onPress={() => {
-                        expandedKeys.includes('schemas')
-                          ? setExpandedKeys(expandedKeys.filter(key => key !== 'schemas'))
-                          : setExpandedKeys([...expandedKeys, 'schemas']);
-                      }}
-                    >
-                      <span className="truncate">Schemas</span>
-                      <Icon icon={expandedKeys.includes('schemas') ? 'minus' : 'plus'} className="text-xs" />
-                    </Button>
-                  </div>
-                  {expandedKeys.includes('schemas') && (
-                    <ListBox
-                      items={Object.entries(schemas).map(([path, item]) => ({
-                        ...item,
-                        id: path,
-                        path,
-                      }))}
-                      onAction={key => navigateToPath(key.toString())}
-                    >
-                      {item => (
-                        <ListBoxItem
-                          className="relative flex h-(--line-height-xs) w-full items-center gap-2 overflow-hidden px-4 text-(--hl) outline-hidden transition-colors select-none hover:bg-(--hl-xs) focus:bg-(--hl-sm)"
-                          id={`components.schemas.${item.path}`}
-                        >
-                          <span className="truncate">{item.path}</span>
-                        </ListBoxItem>
-                      )}
-                    </ListBox>
-                  )}
-                </div>
-              )}
-              {/* Security */}
-              {securitySchemes && (
-                <div className="divide-y divide-solid divide-(--hl-md)">
-                  <div>
-                    <Button
-                      className="flex w-full items-center justify-between gap-2 p-(--padding-sm) text-sm text-(--hl) uppercase select-none hover:bg-(--hl-sm) focus:bg-(--hl-sm)"
-                      onPress={() => {
-                        expandedKeys.includes('security')
-                          ? setExpandedKeys(expandedKeys.filter(key => key !== 'security'))
-                          : setExpandedKeys([...expandedKeys, 'security']);
-                      }}
-                    >
-                      <span className="truncate">Security</span>
-                      <Icon icon={expandedKeys.includes('security') ? 'minus' : 'plus'} className="text-xs" />
-                    </Button>
-                  </div>
-                  {expandedKeys.includes('security') && (
-                    <ListBox
-                      items={Object.entries(securitySchemes).map(([path, item]) => ({
-                        ...item,
-                        id: path,
-                        path,
-                      }))}
-                      onAction={key => navigateToPath(key.toString())}
-                    >
-                      {item => (
-                        <ListBoxItem
-                          className="relative flex h-(--line-height-xs) w-full items-center gap-2 overflow-hidden px-4 text-(--hl) outline-hidden transition-colors select-none hover:bg-(--hl-xs) focus:bg-(--hl-sm)"
-                          id={`components.securitySchemes.${item.path}`}
-                        >
-                          <span className="truncate">{item.path}</span>
-                        </ListBoxItem>
-                      )}
-                    </ListBox>
-                  )}
-                </div>
-              )}
-            </div>
-            {isEnvironmentModalOpen && (
-              <WorkspaceEnvironmentsEditModal onClose={() => setEnvironmentModalOpen(false)} />
-            )}
-            {isCookieModalOpen && <CookiesModal setIsOpen={setIsCookieModalOpen} />}
-            {isCertificatesModalOpen && <CertificatesModal onClose={() => setCertificatesModalOpen(false)} />}
-            {isNewMockServerModalOpen && (
-              <NewWorkspaceModal
-                isOpen={isNewMockServerModalOpen}
-                project={activeProject}
-                storageRules={storageRules}
-                scope="mock-server"
-                sourceApiSpec={apiSpec}
-                source="design-view"
-                onOpenChange={setNewMockServerModalOpen}
-              />
-            )}
-            {isViewRulesetModalOpen && (
-              <ModalOverlay
-                isOpen={isViewRulesetModalOpen}
-                onOpenChange={setIsViewRulesetModalOpen}
-                isDismissable
-                className="fixed top-0 left-0 z-10 flex h-(--visual-viewport-height) w-full items-center justify-center bg-black/30"
-              >
-                <Modal className="max-h-full w-full max-w-[900px] overflow-y-auto rounded-md border border-solid border-(--hl-sm) bg-(--color-bg) p-(--padding-lg) text-(--color-font)">
-                  <Dialog className="relative outline-hidden">
-                    {({ close }) => (
-                      <>
-                        <div className="flex items-center justify-between gap-2 pb-(--padding-sm)">
-                          <Heading slot="title" className="mb-[14px] text-[22px] leading-[34px]">
-                            Existing Ruleset Contents
-                          </Heading>
-                          <Button
-                            aria-label="Close ruleset content viewer"
-                            onPress={close}
-                            className="fa fa-times absolute top-0 right-0 text-xl"
-                          />
-                        </div>
-                        {rulesetContent && (
-                          <CodeEditor
-                            id="ruleset-viewer"
-                            mode="yaml"
-                            readOnly
-                            noLint
-                            defaultValue={rulesetContent}
-                            className="h-[400px]"
-                          />
-                        )}
-                      </>
-                    )}
-                  </Dialog>
-                </Modal>
-              </ModalOverlay>
-            )}
-          </div>
-        </Panel>
-        <PanelResizeHandle className="h-full w-px bg-(--hl-md)" />
-        <Panel id="workspace-content" className="flex flex-col">
-          {specPaneToolbar}
-          <PanelGroup autoSaveId="insomnia-panels" direction={direction} className="min-h-0 flex-1">
-            <Panel id="pane-one" minSize={10} className="pane-one theme--pane">
-              <div className="flex h-full w-full flex-col">
-                <PanelGroup autoSaveId="insomnia-spec-vertical" direction="vertical" className="min-h-0 flex-1">
-                  <Panel id="spec-editor" defaultSize={80} minSize={20} className="relative overflow-hidden">
-                    {specEditor}
-                  </Panel>
-                  {apiSpec.contents && (
-                    <>
-                      <PanelResizeHandle className="h-px w-full bg-(--hl-md)" />
-                      <Panel
-                        ref={lintPanelRef}
-                        id="lint-panel"
-                        defaultSize={0}
-                        minSize={10}
-                        collapsible
-                        onCollapse={() => setIsLintPaneOpen(false)}
-                        onExpand={() => setIsLintPaneOpen(true)}
-                        className="flex flex-col overflow-hidden"
-                      >
-                        {isLintPaneOpen && (
-                          <div data-testid="lint-panel" className="box-border flex h-full flex-col">
-                            {lintToolbar}
-                            {lintMessageList}
-                          </div>
-                        )}
-                      </Panel>
-                    </>
-                  )}
-                </PanelGroup>
-                {apiSpec.contents && !isLintPaneOpen ? lintToolbar : null}
-              </div>
-            </Panel>
-            {isSpecPaneOpen && (
-              <>
-                <PanelResizeHandle
-                  className={direction === 'horizontal' ? 'h-full w-px bg-(--hl-md)' : 'h-px w-full bg-(--hl-md)'}
-                />
-                <Panel id="pane-two" minSize={10} className="pane-two theme--pane">
-                  <SwaggerUIDiv text={apiSpec.contents} />
-                </Panel>
-              </>
-            )}
-          </PanelGroup>
-        </Panel>
-      </PanelGroup>
-    </div>
-  );
-};
-
-export default Component;
